+++ capi/autd3capi/src/modulation/sine.rs
@@ -32,8 +32,11 @@ pub unsafe extern "C" fn AUTDModulationSineWithSamplingConfig(
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDModulationSineWithAmp(m: ModulationPtr, amp: float) -> ModulationPtr {
-    ModulationPtr::new(take_mod!(m, Sine).with_amp(amp))
+pub unsafe extern "C" fn AUTDModulationSineWithIntensity(
+    m: ModulationPtr,
+    intensity: u8,
+) -> ModulationPtr {
+    ModulationPtr::new(take_mod!(m, Sine).with_intensity(intensity))
 }
 
 #[no_mangle]
@@ -49,7 +52,7 @@ pub unsafe extern "C" fn AUTDModulationSineWithPhase(
 #[must_use]
 pub unsafe extern "C" fn AUTDModulationSineWithOffset(
     m: ModulationPtr,
-    offset: float,
+    offset: u8,
 ) -> ModulationPtr {
     ModulationPtr::new(take_mod!(m, Sine).with_offset(offset))
 }
@@ -68,9 +71,9 @@ mod tests {
             let cnt = create_controller();
 
             let m = AUTDModulationSine(150);
-            let m = AUTDModulationSineWithAmp(m, 1.);
+            let m = AUTDModulationSineWithIntensity(m, 255);
             let m = AUTDModulationSineWithPhase(m, 0.);
-            let m = AUTDModulationSineWithOffset(m, 0.5);
+            let m = AUTDModulationSineWithOffset(m, 127);
             let div = 10240;
             let m = AUTDModulationSineWithSamplingConfig(
                 m,